requirements.txt
transformers
torch
torchvision
Pillow